+++ lib/qlevar_router.dart
@@ -1,3 +1,3 @@
-export 'src/pages.dart';
+export 'src/qpages.dart';
 export 'src/qr.dart';
 export 'src/types.dart';